Update Programmatic Transaction Management

+++ s07-spring-transaction/ep05-programmatic-transaction/src/main/java/com/jdc/transaction/repo/MemberRepo.java
@@ -2,6 +2,6 @@ package com.jdc.transaction.repo;
 
 public interface MemberRepo {
 
-	int countById(int memberId);
+	long countById(int memberId);
 
 }
